Reorganize ToastProivider
