+++ packages/toast/src/ToastContext/ToastContext.tsx
@@ -1,9 +1,6 @@
 import React from 'react';
 
-import { ToastContainer } from '../ToastContainer';
-
-import { ToastContextProps, ToastProviderProps } from './ToastContext.types';
-import { useToastReducer } from './ToastReducer';
+import { ToastContextProps } from './ToastContext.types';
 
 const initialToastContextValue = {
   pushToast: () => '',
@@ -17,26 +14,3 @@ const initialToastContextValue = {
 export const ToastContext = React.createContext<ToastContextProps>(
   initialToastContextValue,
 );
-
-/**
- * Toast Provider
- */
-export const ToastProvider = ({
-  children,
-  initialValue,
-  portalClassName,
-}: React.PropsWithChildren<ToastProviderProps>) => {
-  const { stack, ...toastFns } = useToastReducer(initialValue);
-
-  return (
-    <ToastContext.Provider
-      value={{
-        ...toastFns,
-        getStack: () => stack,
-      }}
-    >
-      {children}
-      <ToastContainer stack={stack} portalClassName={portalClassName} />
-    </ToastContext.Provider>
-  );
-};
