+++ packages/toast/src/ToastContext/ToastProvider/index.ts
@@ -1,0 +1,1 @@
+export { ToastProvider } from './ToastProvider';
